% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_methQTL.R
\name{call.methQTL.block}
\alias{call.methQTL.block}
\title{call.methQTL.block}
\usage{
call.methQTL.block(cor.block, meth.data, geno.data, covs,
  model.type = qtl.getOption("linear.model.type"))
}
\arguments{
\item{cor.block}{A single correlation block as determined by \code{\link{compute.correlation.blocks}}}

\item{meth.data}{The methylation data matrix}

\item{geno.data}{The genotype data matrix}

\item{covs}{A \code{data.frame} specifying the covariates used for the analysis}

\item{model.type}{Type of the linear model to be used. Supported options are:
\describe{
  \item{categorical.anova}{Performs linear regression after transforming the genotype states (0=homozygous
  reference allele, 1=heteroygous, 2=homoygous alternate allele) into categorical variables and then perform
  ANOVA}
  \item{classical.linear}{Performs linear regression with keeping the genotype calls as numeric values. This
  induces an ordering between heterozygous and homozygous reference. We then use the p-value from the t-statistic
  and the associated beta value.}
}}
}
\value{
A vector of three elements:
   \describe{
     \item{1}{The p-value of the best methQTL}
     \item{2}{The estimated coefficient of the linear model}
     \item{3}{The position of the SNP}
     \item{4}{The position of the CpG}
   }
}
\description{
This function computes a methQTL per correlation block.
}
\details{
methQTL are called using a linear modeling strategy: The DNA methylation state of the CpG is considered
    the output, while the SNP genotype and the potential covariates are the input. We compute a linear model
    for each CpG in the correlation block and each SNP. We then select the best methQTL according to the lowest
    p-value and return it.
}
\author{
Michael Scherer
}
