% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_methQTL.R
\name{callMethQTLBlock}
\alias{callMethQTLBlock}
\title{callMethQTLBlock}
\usage{
callMethQTLBlock(
  cor.block,
  meth.data,
  geno.data,
  covs,
  anno.meth,
  anno.geno,
  model.type = qtlGetOption("linear.model.type"),
  n.permutations = qtlGetOption("n.permutations")
)
}
\arguments{
\item{cor.block}{A single correlation block as determined by \code{\link{computeCorrelationBlocks}}}

\item{meth.data}{The methylation data matrix}

\item{geno.data}{The genotype data matrix}

\item{covs}{A \code{data.frame} specifying the covariates used for the analysis}

\item{anno.meth}{A \code{data.frame} containing genomic annotations of the CpGs}

\item{anno.geno}{A \code{data.frame} containing genomic annotations for the SNPs}

\item{model.type}{Type of the linear model to be used. Supported options are:
\describe{
  \item{categorical.anova}{Performs linear regression after transforming the genotype states (0=homozygous
  reference allele, 1=heteroygous, 2=homoygous alternate allele) into categorical variables and then perform
  ANOVA}
  \item{classical.linear}{Performs linear regression with keeping the genotype calls as numeric values. This
  induces an ordering between heterozygous and homozygous reference. We then use the p-value from the t-statistic
  and the associated beta value.}
}}

\item{n.permutations}{Number of permutations used to correct the p-values. Only has an influence, if the parameter
\code{"p.value.correction"="permutation".}}
}
\value{
A vector of three elements:
   \describe{
     \item{1}{The p-value of the best methQTL}
     \item{2}{The estimated coefficient of the linear model}
     \item{3}{The position of the SNP}
     \item{4}{The position of the CpG}
   }
}
\description{
This function computes a methQTL per correlation block.
}
\details{
methQTL are called using a linear modeling strategy: The DNA methylation state of the CpG is considered
    the output, while the SNP genotype and the potential covariates are the input. We compute a linear model
    for each CpG in the correlation block and each SNP. We then select the best methQTL according to the lowest
    p-value and return it.
}
\author{
Michael Scherer
}
