% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{do.imputation}
\alias{do.imputation}
\title{do.imputation}
\usage{
do.imputation(bed.file, bim.file, fam.file, out.dir)
}
\arguments{
\item{bed.file}{Path to the PLINK BED file}

\item{bim.file}{Path to the PLINK BIM file}

\item{fam.file}{Path to the PLINK FAM file}

\item{out.dir}{The output directory}
}
\value{
A vector with three elements: \describe{
  \item{\code{"bed.file"}}{Path to the BED file for PLINK}
  \item{\code{"bim.file"}}{Path to the BIM file for PLINK}
  \item{\code{"fam.file"}}{Path to the FAM file for PLINK}
}
}
\description{
Function to perform imputation from PLINK data
}
\details{
Since PLINK files are converted to VCF files back and forth, unserscores are not allowed
   in the sample IDs in case imputation is performed.
}
\author{
Michael Scherer
}
