% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_methQTL.R
\name{do.methQTL.chromosome}
\alias{do.methQTL.chromosome}
\title{do.methQTL.chromosome}
\usage{
do.methQTL.chromosome(meth.qtl, chrom, sel.covariates, p.val.cutoff,
  out.dir = NULL)
}
\arguments{
\item{meth.qtl}{An Object of type \code{\link{methQTLInput-class}}.}

\item{chrom}{Character vector represeting the chromosome to be investigated.}

\item{sel.covariates}{Covariates as column names of the sample annotation sheet stored in \code{meth.qtl} to be
used for covariate adjustment.}

\item{p.val.cutoff}{The p-value used for methQTL calling}

\item{out.dir}{Optional argument specifying the output directory}
}
\value{
A data frame with seven columns:
       \describe{
         \item{CpGs}{The CpG ID chosen to be the representative CpG in the methQTL}
         \item{SNP}{The SNP ID (as rsNNNNNN) involved in the methQTL}
         \item{Beta}{The coefficient estimate of the linear model}
         \item{P.value}{The p-value associated with the coefficient estimate}
         \item{Chromosome}{The chromosome name}
         \item{Position.CpG}{The genomic position of the CpG}
         \item{Position.SNP}{The genomic position of the SNP}
         \item{Distance}{The distance between the CpG and the SNP}
       }
}
\description{
This functions computes the methQTL interactions for a single chromosome
}
\author{
Michael Scherer
}
