% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_methQTL.R
\name{do.methQTL}
\alias{do.methQTL}
\title{do.methQTL}
\usage{
do.methQTL(meth.qtl, sel.covariates = NULL, p.val.cutoff = 1e-05,
  ncores = 1)
}
\arguments{
\item{meth.qtl}{An object of type \code{\link{methQTLInput-class}} on which methQTL computation is to be performed}

\item{sel.covariates}{Covariates as column names of the sample annotation sheet stored in \code{meth.qtl} to be
used for covariate adjustment.}

\item{p.val.cutoff}{The p-value cutoff used for methQTL calling}

\item{ncores}{The number of cores used.}
}
\value{
An object of type \code{\link{methQTLResult-class}} containing the called methQTL interactions.
}
\description{
Function to compute methQTL given DNA methylation and genotyping data.
}
\details{
The process is split into 4 steps:
         \describe{
           \item{1}{First the two matrices are split according to the chromosomes.}
           \item{2}{We then compute correlations among the CpGs and compute CpG correlation blocks.}
           \item{3}{In each of the CpG correlation blocks, simple linear models with the CpG methylation state
             of all CpGs as output and the SNP genotype state and all possible covariates as input are computed.}
           \item{4}{For each of the CpG correlation blocks, the CpGs with the lowest p-value is selected as the
             representative methQTL per block.}
         }
}
\author{
Michael Scherer
}
