% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{doGenoImport}
\alias{doGenoImport}
\title{doGenoImport}
\usage{
doGenoImport(data.location, s.anno, s.id.col, out.folder, ...)
}
\arguments{
\item{data.location}{Path to the directory, where the PLINK files are stored. For more information, see \code{\link{doImport}}}

\item{s.anno}{The sample annotation sheet.}

\item{s.id.col}{The column name of the sample annotation sheet specifying the sample identifiers.}

\item{out.folder}{The output folder for storing diagnostic plots.}

\item{...}{Futher parameters passed to \code{\link{doGenoImportImputed}} or \code{\link{doGenoImportIDAT}} depending on the
option \code{'geno.data.type'}.}
}
\value{
A list of three elements:
       \describe{
         \item{data}{The processed genotyping data as a data.frame}
         \item{annotation}{The genomic annotation of the SNPs in \code{data}}
         \item{pheno.data}{The, potentially modified, sample annotation sheet}
       }
}
\description{
This routine performs data import and processing of genotyping data using the \code{PLINK} program.
}
\details{
This function should seldomly be called individually, but rather through \code{\link{doImport}}
}
\seealso{
doImport
}
\author{
Michael Scherer
}
