% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{doGenoImportImputed}
\alias{doGenoImportImputed}
\title{doGenoImportImputed}
\usage{
doGenoImportImputed(
  dos.file,
  id.map,
  s.anno,
  s.id.col,
  out.folder,
  tab.sep = " "
)
}
\arguments{
\item{dos.file}{A dosage file containing the dosage information for the imputed data. Can be gzipped.}

\item{id.map}{A path to a file mapping the row names of the dosage file to SNP (rs) identifiers.}

\item{s.anno}{The sample annotation sheet as a data frame.}

\item{s.id.col}{A column name in the sample annotation sheet specifying the sample identifers.}

\item{out.folder}{A path to a folder to which the results are to be stored.}

\item{tab.sep}{The table separator used for \code{dos.file} and \code{id.map}.}
}
\value{
A list of three elements:
       \describe{
         \item{data}{The processed genotyping data as a data.frame}
         \item{annotation}{The genomic annotation of the SNPs in \code{data}}
         \item{pheno.data}{The, potentially modified, sample annotation sheet}
       }
}
\description{
This function executes import of imputed genotyping data
}
\details{
Genotyping data can also be imported from imputed genotype data. However, data needs to be already
 preprocessed, such that for instance SNPs outside of the Hardy-Weinberg equilibrium, sites with too many
 missing values or SNPs with a minimum minor allele frequency are removed/kept, since the data cannot be further
 processed by \code{plink}. We expect as input a tabular file that contains SNPs in the rows and samples in the
 columns, where each entry is a continous numeric value containing the dosage.
}
\author{
Michael Scherer
}
