% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{doMethImport}
\alias{doMethImport}
\title{doMethImport}
\usage{
doMethImport(
  data.location,
  assembly = "hg19",
  s.anno,
  s.id.col,
  tab.sep = ",",
  snp.location = NULL,
  out.folder = NULL,
  ...
)
}
\arguments{
\item{data.location}{The location of the data files. See \code{\link{doImport}} for further details.}

\item{assembly}{The assembly used.}

\item{s.anno}{Path to the sample annotation sheet}

\item{s.id.col}{Column name in the sample annotation sheet specifying the sample identifiers.}

\item{tab.sep}{Table separator used.}

\item{snp.location}{Locations of SNPs called in the genotyped processing to be removed from the list of CpGs.}

\item{out.folder}{If not \code{NULL} a directory in which intermediate results are to be written}

\item{...}{Further parameters passed to, e.g., \code{\link{qtlRunSegmentation}}, \code{\link{doGenoImportIDAT}}}
}
\value{
A list with five elements:
         \describe{
           \item{sample}{The samples used in the dataset as a character vector}
           \item{data}{The DNA methylation data matrix as the results of import and preprocessing}
           \item{annotation}{The genomic annotation of the sites present in \code{data}}
           \item{platform}{The platform used, e.g. \code{'probesEPIC'}}
           \item{segmentation}{If segmentation was performed, the segmentation as a \code{GRanges} object}
         }
}
\description{
This function performs import of DNA methylation data and basic preprocessing steps implemented in the \code{\link{RnBeads}} package.
}
\details{
This function execute the import and preprocessing modules of RnBeads. In the default setting, a common
         option setting for Illumina BeadArray data is used (described in \code{inst/extdata/rnbeads_options.xml}).
         For further option setting, you should use the \code{rnbeads.options} option of the methQTL package to
         specify another XML file with custom RnBeads options.
         This function should seldomly be called individually, but rather through \code{\link{doImport}}
}
\seealso{
doImport
}
\author{
Michael Scherer
}
