% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLResult-class.R
\docType{methods}
\name{filter.pval,methQTLResult-method}
\alias{filter.pval,methQTLResult-method}
\title{filter.pval}
\usage{
\S4method{filter.pval}{methQTLResult}(object, p.val.cutoff = 0.01)
}
\arguments{
\item{object}{The \code{\link{methQTLResult-class}} object to be filtered}

\item{p.val.cutoff}{The p-value cutoff to be employed}
}
\value{
The filtered \code{\link{methQTLResult-class}} object
}
\description{
This functions filters the methQTL results according to a given p-value cutoff
}
\author{
Michael Scherer
}
