% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQTLResult-class.R
\docType{methods}
\name{filterPval,MethQTLResult-method}
\alias{filterPval,MethQTLResult-method}
\alias{filterPval}
\title{filterPval}
\usage{
\S4method{filterPval}{MethQTLResult}(object, p.val.cutoff = 0.01)
}
\arguments{
\item{object}{The \code{\link{MethQTLResult-class}} object to be filtered}

\item{p.val.cutoff}{The p-value cutoff to be employed}
}
\value{
The filtered \code{\link{MethQTLResult-class}} object
}
\description{
This functions filters the methQTL results according to a given p-value cutoff
}
\examples{
meth.qtl.res <- loadMethQTLResult(system.file("extdata","MethQTLResult_chr18",package="MAGAR"))
meth.qtl.res <- filterPval(meth.qtl.res)
meth.qtl.res
}
\author{
Michael Scherer
}
