% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate.fastQTL.input}
\alias{generate.fastQTL.input}
\title{generate.fastQTL.input}
\usage{
generate.fastQTL.input(
  meth.qtl,
  chrom,
  correlation.block,
  sel.covariates,
  out.dir
)
}
\arguments{
\item{meth.qtl}{An object of type \code{\link{methQTLInput-class}} with methylation and genotyping information}

\item{chrom}{The chromosome to be investigated.}

\item{correlation.block}{The correlation blocks generated.}

\item{sel.covariates}{The selected covariates as a \code{data.frame}}

\item{out.dir}{The target output directory}
}
\value{
A vector comprising the following elements:\describe{
          \item{\code{genotypes:}}{The path to the genotypes file}
          \item{\code{phenotypes:}}{The path to the DNA methylation data file}
          \item{\code{covariates:}}{The path to the covariates file}
}
}
\description{
This function generates the required input to FastQTL and stores the files on disk.
}
