% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get.overlap.universe}
\alias{get.overlap.universe}
\title{get.overlap.universe}
\usage{
get.overlap.universe(meth.qtl.res, type)
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{methQTLResult-class}} or a list of such objects}

\item{type}{The type of annotation to be overlapped. Needs to be \code{'SNP'}, \code{'CpG'} or \code{'cor.block'}}
}
\value{
A list with two GRanges objects, one containing the overlapped set and the other the
   union of input data points as elements \code{'all.qtl'} and \code{'all.input'}
}
\description{
This function overlaps results from a list of methQTLResults and returns the union of all
the input data points used.
}
\author{
Michael Scherer
}
