% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLResult-class.R
\docType{methods}
\name{getCorrelationBlocks,methQTLResult-method}
\alias{getCorrelationBlocks,methQTLResult-method}
\alias{getCorrelationBlocks}
\title{getCorrelationBlocks}
\usage{
\S4method{getCorrelationBlocks}{methQTLResult}(object)
}
\arguments{
\item{object}{An object of class \code{\link{methQTLResult-class}}.}
}
\value{
A \code{list} object containing the correlation blocks.


}
\description{
Returns the correlation blocks defined for the given dataset
}
