% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLInput-class.R
\docType{methods}
\name{getPheno,methQTLInput-method}
\alias{getPheno,methQTLInput-method}
\alias{getPheno,methQTL-method}
\alias{getPheno}
\title{getPheno}
\usage{
\S4method{getPheno}{methQTLInput}(object)
}
\arguments{
\item{object}{An object of class \code{\link{methQTLInput-class}}.}
}
\value{
The phenotypic data either as a \code{data.frame}.
}
\description{
Returns phenotypic information for the given dataset.
}
\examples{
{
meth.qtl <- loadMethQTL(system.file("extdata/reduced_methQTL",package="MAGAR"))
head(getPheno(meth.qtl))
}
}
