% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLResult-class.R
\docType{methods}
\name{getResult.GWASMap,methQTLResult-method}
\alias{getResult.GWASMap,methQTLResult-method}
\title{getResult.GWASMap}
\usage{
\S4method{getResult.GWASMap}{methQTLResult}(object, meth.qtl)
}
\arguments{
\item{object}{An of type \code{\link{methQTLResult-class}}.}

\item{meth.qtl}{An object of type \code{\link{methQTLInput-class}} containing further information about the QTLs}
}
\value{
The methQTL results as a \code{data.frame} with each row being a methQTL.
}
\description{
Returns the methQTL results in the format used as input to GWAS-map and stores in the object.
}
