% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQTLInput-class.R
\docType{methods}
\name{getSamples,MethQTLInput-method}
\alias{getSamples,MethQTLInput-method}
\alias{getSamples,methQTL-method}
\alias{getSamples}
\title{getSamples}
\usage{
\S4method{getSamples}{MethQTLInput}(object)
}
\arguments{
\item{object}{An object of class \code{\link{MethQTLInput-class}}.}
}
\value{
The samples of the dataset as a character vector.
}
\description{
Returns the samples of the given dataset.
}
\examples{
meth.qtl <- loadMethQTLInput(system.file("extdata","reduced_methQTL",package="MAGAR"))
getSamples(meth.qtl)
}
