% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLInput-class.R
\docType{methods}
\name{impute.meth,methQTLInput-method}
\alias{impute.meth,methQTLInput-method}
\alias{impute.meth,methQTL-method}
\title{impute.meth}
\usage{
\S4method{impute.meth}{methQTLInput}(object)
}
\arguments{
\item{object}{An object of class \code{\link{methQTLInput-class}}.}
}
\value{
The object with imputed values.
}
\description{
Replaces missing values in the DNA methylation data matrix by imputed values
}
