% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQTLInput-class.R
\docType{methods}
\name{imputeMeth,MethQTLInput-method}
\alias{imputeMeth,MethQTLInput-method}
\alias{imputeMeth,methQTL-method}
\alias{imputeMeth}
\title{imputeMeth}
\usage{
\S4method{imputeMeth}{MethQTLInput}(object)
}
\arguments{
\item{object}{An object of class \code{\link{MethQTLInput-class}}.}
}
\value{
The object with imputed values.
}
\description{
Replaces missing values in the DNA methylation data matrix by imputed values
}
\examples{
meth.qtl <- loadMethQTLInput(system.file("extdata","reduced_methQTL",package="MAGAR"))
meth.qtl.imp <- imputeMeth(meth.qtl)
}
