% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLResult-class.R
\name{join.methQTL}
\alias{join.methQTL}
\title{join.methQTLResult}
\usage{
join.methQTL(obj.list)
}
\arguments{
\item{obj.list}{A list of \code{\link{methQTLResult-class}} objects to be joined}
}
\value{
An object of type \code{\link{methQTLResult-class}} containing the combined information
}
\description{
This function combines a list of \code{\link{methQTLResult-class}} objects.
}
\author{
Michael Scherer
}
