% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLInput-class.R
\name{load.methQTL}
\alias{load.methQTL}
\title{load.methQTL}
\usage{
load.methQTL(path)
}
\arguments{
\item{path}{Path to the directory that has been created by \code{\link{save.methQTL,methQTLInput-method}}.}
}
\value{
The object of type \code{\link{methQTLInput-class}} that has been stored on disk.
}
\description{
This functions load a \code{\link{methQTLInput-class}} object from disk.
}
\author{
Michael Scherer
}
