% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQTLInput-class.R
\name{loadMethQTLInput}
\alias{loadMethQTLInput}
\title{loadMethQTLInput}
\usage{
loadMethQTLInput(path)
}
\arguments{
\item{path}{Path to the directory that has been created by \code{\link{saveMethQTLInput,MethQTLInput-method}}.}
}
\value{
The object of type \code{\link{MethQTLInput-class}} that has been stored on disk.
}
\description{
This functions load a \code{\link{MethQTLInput-class}} object from disk.
}
\examples{
meth.qtl <- loadMethQTLInput(system.file("extdata","reduced_methQTL",package="MAGAR"))
meth.qtl
}
\author{
Michael Scherer
}
