% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQTLResult-class.R
\name{loadMethQTLResult}
\alias{loadMethQTLResult}
\title{loadMethQTLResult}
\usage{
loadMethQTLResult(path)
}
\arguments{
\item{path}{Path to the directory that has been created by \code{saveMethQTLResult,MethQTLResult-method}.}
}
\value{
The object of type \code{\link{MethQTLResult-class}} that has been stored on disk.
}
\description{
This functions load a \code{\link{MethQTLResult-class}} object from disk.
}
\examples{
meth.qtl.res <- loadMethQTLResult(system.file("extdata","MethQTLResult_chr18",package="MAGAR"))
meth.qtl.res
}
\author{
Michael Scherer
}
