% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{overlapQTLs}
\alias{overlapQTLs}
\title{overlapQTLs}
\usage{
overlapQTLs(meth.qtl.result.list, type)
}
\arguments{
\item{meth.qtl.result.list}{A named list with each entry being an object of type \code{\link{methQTLResult-class}}.
The names are used in the visualization.}

\item{type}{Determines if either the SNP (default), the CpG, or the correlation block
\code{'cor.block'} is to be visualized}
}
\value{
A list with \code{length(meth.qtl.result.list)} elements, containing IDs of methQTL
    interactions according to the option \code{type}.
}
\description{
This function overlaps a list of methQTLs to determine which interactions are common.
}
\author{
Michael Scherer
}
