% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAPMap.R
\name{parseGWASMapOutput}
\alias{parseGWASMapOutput}
\title{parseGWASMapOutput}
\usage{
parseGWASMapOutput(file, sep = ",", dec = ".")
}
\arguments{
\item{file}{The path to the GWASMap CSV file}

\item{sep}{The table separator}

\item{dec}{The decimal point separator}
}
\value{
A list object with each element being the first line of the GWASMap output file.
     Each of the distinct elements in the first column is used as a category and the associated
     table is parsed separately as a table.
}
\description{
This method takes as input a path to a CSV file containing the GWASMap output
}
\author{
Michael Scherer
}
