% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.SNP.CpG.interaction}
\alias{plot.SNP.CpG.interaction}
\title{plot.SNP.CpG.interaction}
\usage{
plot.SNP.CpG.interaction(meth.qtl, cpg, snp, out.dir = NULL)
}
\arguments{
\item{meth.qtl}{An object of type \code{\link{methQTLInput-class}} containing the methylation and genotype information
for the given CpG and the given SNP}

\item{cpg}{The CpG identifier as a character (e.g. cg12345678)}

\item{snp}{The SNP identifier as a character (e.g. rs12345678)}

\item{out.dir}{If specified, the plot is stored as a pdf in this directory}
}
\value{
An object of type \code{ggplot} comparing the CpG methylation states as boxplots across the different genotype states
}
\description{
Compares the methylation states of a given CpG for the genotype states availabe at the given SNP
}
\author{
Michael Scherer
}
