% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{qtl.annotation.enrichment}
\alias{qtl.annotation.enrichment}
\title{qtl.annotation.enrichment}
\usage{
qtl.annotation.enrichment(
  meth.qtl.res,
  type = "SNP",
  annotation = "cpgislands"
)
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{methQTLResult}} or a list of such objects.}

\item{type}{The type of methQTL to be visualized. Can be either \code{'SNP'}, \code{'CpG'},
or \code{'cor.block'}}

\item{annotation}{The genomic annotation to be used. Can be the ones available in \code{\link{region.types()}}}
}
\value{
The p-value of the Fisher exact test for enrichment
}
\description{
This functions performs enrichment analysis using the Fisher's test for the methQTLs detected
with respect to different genomic annotations.
}
\details{
We use all data points that have been used to calculate methQTLs as the background
   and compare the overlaps with the annotation of interest in comparison to the methQTLs that
   have been computed.
}
\author{
Michael Scherer
}
