% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtl.correlate.cor.block.stat}
\alias{qtl.correlate.cor.block.stat}
\title{qtl.correlate.cor.block.stat}
\usage{
qtl.correlate.cor.block.stat(
  meth.qtl.res,
  meth.qtl,
  stat = "p.val.adj.fdr",
  size.type = "num.CpGs"
)
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{methQTLResult-class}} containing methQTLs and correlation blocks.}

\item{meth.qtl}{An object of type \code{\link{meth.qtl}} containing CpG annotations.}

\item{stat}{The statistic that is to be compared. Can be one of \code{'P.value'}, \code{'Beta'},
\code{'Distance'}, or \code{'p.val.adj.fdr'}(default).}

\item{size.type}{The size of correlation block to be used. Can be either \code{'num.CpGs'} (default)
or \code{'genomic'}.}
}
\value{
A scatterplot and associated correlations as an objec to type \code{ggplot}
}
\description{
This function correlates the size of the correlation block, a particular CpG is part of, to the statistic
that has been found for a methQTL this CpG is involved in.
}
\author{
Michael Scherer
}
