% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{qtl.getOption}
\alias{qtl.getOption}
\title{qtl.getOption
Print the value of the global option}
\usage{
qtl.getOption(names)
}
\arguments{
\item{names}{string or character vector containing the names of the options to be printed}
}
\value{
the option for the specified option
}
\description{
qtl.getOption
Print the value of the global option
}
\author{
Michael Scherer
}
