% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{qtl.json2options}
\alias{qtl.json2options}
\title{qtl.json2options}
\usage{
qtl.json2options(path)
}
\arguments{
\item{path}{Path to a JSON file containing the options to be specified}
}
\description{
This function reads an option setting from a JSON file and applies them to the current session
}
\author{
Michael Scherer
}
