% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{qtl.lola.enrichment}
\alias{qtl.lola.enrichment}
\title{qtl.lola.enrichment}
\usage{
qtl.lola.enrichment(meth.qtl.result.list, type = "SNP", lola.db = NULL)
}
\arguments{
\item{meth.qtl.result.list}{A list containing objects of type \code{\link{methQTLResult-class}}}

\item{type}{The type of methQTL to be visualized. Can be either \code{'SNP'}, \code{'CpG'},
or \code{'cor.block'}}

\item{lola.db}{The location of a LOLA DB already downloaded}
}
\value{
The LOLA enrichment result
}
\description{
This functions performn LOLA enrichment analysis for the sites that are shared across all
methQTLs in the input list.
}
\author{
Michael Scherer
}
