% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtl.lola.plot}
\alias{qtl.lola.plot}
\title{qtl.lola.plot}
\usage{
qtl.lola.plot(
  meth.qtl.res,
  type = "SNP",
  lola.db = NULL,
  assembly = "hg19",
  pvalCut = 0.01
)
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{methQTLResult-class}} or a list of such objects}

\item{type}{The type of methQTL to be visualized. Can be either \code{'SNP'}, \code{'CpG'},
or \code{'cor.block'}}

\item{lola.db}{The location of a LOLA DB already downloaded}

\item{assembly}{The assembly used (defaul \code{'hg19'})}

\item{pvalCut}{The p-value cutoff employed}
}
\value{
The LOLA enrichment bar plot as a \code{ggplot} object
}
\description{
This function plots the LOLA enrichment results using the \code{\link{lolaBarPlot}} routine.
}
\author{
Michael Scherer
}
