% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{QTL.OPTIONS}
\alias{QTL.OPTIONS}
\title{options.R}
\format{An object of class \code{environment} of length 29.}
\usage{
QTL.OPTIONS
}
\description{
This files contains code to generate the options of the methQTL package.
}
\keyword{datasets}
