% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{qtl.options2json}
\alias{qtl.options2json}
\title{qtl.options2json}
\usage{
qtl.options2json(path = file.path(getwd(), "methQTL_options.json"))
}
\arguments{
\item{path}{A filename, to which the option setting is to be saved}
}
\description{
This function stores the current options setting as a JSON file at the specified path
}
\author{
Michael Scherer
}
