% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtl.plot.annotation.enrichment}
\alias{qtl.plot.annotation.enrichment}
\title{qtl.plot.annotation.enrichment}
\usage{
qtl.plot.annotation.enrichment(meth.qtl.res, ...)
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{methQTLResult-class}} or a list of such objects.}

\item{...}{Further parameters passed to \code{\link{qtl.annotation.enrichment }}}
}
\description{
This functions returns and enrichment plot for different genomic annotation enrichments
}
\seealso{
qtl.annotation.enrichment
}
\author{
Michael Scherer
}
