% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtl.plot.base.substitution}
\alias{qtl.plot.base.substitution}
\title{qtl.plot.base.substitution}
\usage{
qtl.plot.base.substitution(meth.qtl.res, ...)
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{methQTLResult-class}} or a list of such objects.}

\item{...}{Further parameters passed to \code{\link{qtl.base.substitution.enrichment}}}
}
\description{
This function returns an enrichment plot for the different base substitutions.
}
\seealso{
qtl.base.substitution.enrichment
}
\author{
Michael Scherer
}
