% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtl.plot.cluster.size}
\alias{qtl.plot.cluster.size}
\title{qtl.plot.cluster.size}
\usage{
qtl.plot.cluster.size(meth.qtl.res, type = "count")
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{methQTLResult-class}}}

\item{type}{Either \code{"genomic"} or \code{"count"}, for genomic size of the correlation
block in base pairs or as the number of CpGs}
}
\value{
An objec of type ggplot containing the histogram as a plot
}
\description{
This functions returns a histogram comprising the (genomic) sizes of the correlation blocks
in the given objet.
}
\author{
Michael Scherer
}
