% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtl.plot.SNP.correlation.block}
\alias{qtl.plot.SNP.correlation.block}
\title{qtl.plot.SNP.correlation.block}
\usage{
qtl.plot.SNP.correlation.block(meth.qtl.res, meth.qtl, snp = NULL)
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{methQTLResult}} containing the methQTL interactions
and the CpG correlation blocks.}

\item{meth.qtl}{An object of type \code{\link{methQTLInput}} containing genotyping, methylation and
annotation information.}

\item{snp}{The SNP identifier, for which the methQTL interaction is to be visualized.}
}
\value{
The plot as an \code{ggplot} object, containing a facet with a scatterplot between the CpG
   methylation state and the SNP dosage. Discrete genotypes are currently not supported.
}
\description{
This functions creates a multi-facet plot with a panel for each CpG in the correlation block that has
a methQTL interaction with the SNP of interest.
}
\author{
Michael Scherer
}
