% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{qtl.run.segmentation}
\alias{qtl.run.segmentation}
\title{qtl.run.segmentation}
\usage{
qtl.run.segmentation(rnb.set, out.folder, train.chr = "chr2")
}
\arguments{
\item{rnb.set}{An object of type \code{\link{RnBSet-class}} with required DNA methylation information}

\item{out.folder}{The output folder to store intermediate results}
}
\value{
A \code{GRanges} object with the segmentation performed
}
\description{
This function performs DNA methylation based segmentation using the 'epicPMDdetect' package
}
\details{
The 'epicPMDdetect' package has been created by Malte Gross
}
\author{
Michael Scherer
}
