% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{qtl.setOption}
\alias{qtl.setOption}
\title{qtl.setOption}
\usage{

  qtl.setOption(rnbeads.options = system.file("inst/extdata/rnbeads_options.xml",
  package = "methQTL"), rnbeads.report = "temp", rnbeads.qc = F,
  HDF5dump = F, hardy.weinberg.p = 0.001,
  minor.allele.frequency = 0.05, missing.values.samples = 0.05)
}
\arguments{
\item{rnbeads.options}{Path to an XML file specifying the RnBeads options used for data import. The default options
are suitable for Illumina Beads Array data sets.}

\item{rnbeads.report}{Path to an existing directory, in which the preprocessing report of RnBeads is to be stored.
Defaults to the temporary file.}

\item{HDF5dump}{Flag indicating, if large matrices are to be stored on disk rather than in main memory using the
\code{\link{HDF5Array}} package.}

\item{hardy.weinberg.p}{P-value used for the markers to be excluded if the do not follow the
Hardy-Weinberg equilibrium as implemented in \code{PLINK}.}

\item{minor.allele.frequency}{Threshold for the minor allele frequency of the SNPs to be used in the analysis.}

\item{missing.values.samples}{Threshold specifying how much missing values per SNP are allowed across the samples
to be included in the analyis.}
}
\description{
Change global options for methQTL calculation
}
\examples{
\donttest{
qtl.getOption("rnbeads.report")
qtl.setOption(rnbeads.report=getwd())
qtl.getOption("rnbeads.report")
}
}
\author{
Michael Scherer
}
