% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtl.upset.plot}
\alias{qtl.upset.plot}
\title{qtl.upset.plot}
\usage{
qtl.upset.plot(meth.qtl.result.list, type = "SNP", ...)
}
\arguments{
\item{meth.qtl.result.list}{A named list with each entry being an object of type \code{\link{methQTLResult-class}}.
The names are used in the visualization.}

\item{type}{Determines if either the SNP (default), the CpG, or the correlation block
\code{'cor.block'} is to be visualized}

\item{...}{Further argument passed to \code{\link[UpSetR]{upset}}}
}
\description{
This function creates an UpSet plot from the given methQTL results
}
\details{
The plot is directly drawn and can be stored on disk using the known R graphic devices
}
\author{
Michael Scherer
}
