% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtl.venn.plot}
\alias{qtl.venn.plot}
\title{qtl.venn.plot}
\usage{
qtl.venn.plot(meth.qtl.result.list, type = "CpG", out.folder,
  out.name = NULL, ...)
}
\arguments{
\item{meth.qtl.result.list}{A named list with each entry being an object of type \code{\link{methQTLResult-class}}.
The names are used in the visualization.}

\item{type}{Determines if either the CpG (default) or the SNP is to be visualized}

\item{out.folder}{Required argument specifying the location to store the resulting plot}

\item{out.name}{Optional argument for the name of the plot on disk (ending needs to be .png)}

\item{...}{Further argument passed to \code{\link[VennDiagram]{venn.diagram}}}
}
\value{
The venn plot object
}
\description{
This function creates a venn plot from a list of methQTL results, showing the overlap between the interactions
}
\author{
Michael Scherer
}
