% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{qtlGetOption}
\alias{qtlGetOption}
\title{qtlGetOption
Print the value of the global option}
\usage{
qtlGetOption(names)
}
\arguments{
\item{names}{string or character vector containing the names of the options to be printed. All options are listed in \code{\link{qtlSetOption}}}
}
\value{
the option for the specified option
}
\description{
qtlGetOption
Print the value of the global option
}
\examples{
{
qtlGetOption("cluster.cor.threshold")
}
}
\author{
Michael Scherer
}
