% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{qtlJSON2options}
\alias{qtlJSON2options}
\title{qtlJSON2options}
\usage{
qtlJSON2options(path)
}
\arguments{
\item{path}{Path to a JSON file containing the options to be specified}
}
\value{
None
}
\description{
This function reads an option setting from a JSON file and applies them to the current session
}
\examples{
qtlJSON2options(system.file("extdata","qtl_options_probesEPIC.json",package="MAGAR"))
}
\author{
Michael Scherer
}
