% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{qtlLOLAEnrichment}
\alias{qtlLOLAEnrichment}
\title{qtlLOLAEnrichment}
\usage{
qtlLOLAEnrichment(
  meth.qtl.res,
  type = "SNP",
  lola.db = NULL,
  assembly = "hg19"
)
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{methQTLResult-class}} or a list of such objects}

\item{type}{The type of methQTL to be visualized. Can be either \code{'SNP'}, \code{'CpG'},
or \code{'cor.block'}}

\item{lola.db}{The location of a LOLA DB already downloaded}

\item{assembly}{The assembly used (defaul \code{'hg19'})}
}
\value{
The LOLA enrichment result
}
\description{
This functions performn LOLA enrichment analysis for the methQTL sites or the sites that are
shared across all methQTLs in the input list.
}
\details{
We use all data points that have been used to calculate methQTLs as the background
 and compare the overlaps with the annotation of interest in comparison to the methQTLs that
 have been computed in case a \code{\link{methQTLResult-class}} is provided. If a list of \code{\link{methQTLResult-class}} objects
 is provided, the intersection between the methQTLs from all objects in the list is compared with the union of all interactions
 that have been tested.
}
\examples{
\donttest{
 meth.qtl.res <- loadMethQTLResult(system.file("extdata/methQTLResult_chr18",package="MAGAR"))
 res <- qtlLOLAEnrichment(meth.qtl.res)
}
}
\author{
Michael Scherer
}
