% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtlManhattanPlot}
\alias{qtlManhattanPlot}
\title{qtlManhattanPlot}
\usage{
qtlManhattanPlot(meth.qtl.result, type = "CpG", stat = "p.val.adj.fdr")
}
\arguments{
\item{meth.qtl.result}{An object of type \code{\link{methQTLResult-class}} containing the methQTL}

\item{type}{Determines if either the CpG (default) or the SNP is to be visualized}

\item{stat}{Determines the statistic that is to be visualized. Can be either \code{P.value}, \code{Beta} or
\code{p.val.adj.fdr}}
}
\value{
None
}
\description{
This function creates a manhattan plot for the given methQTL result
}
\details{
A plot is shown that contains chromosome-wise interactions.
}
\examples{
meth.qtl.res <- loadMethQTLResult(system.file("extdata","methQTLResult_chr18",package="MAGAR"))
qtlManhattanPlot(meth.qtl.res)
}
\author{
Michael Scherer
}
