% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{qtlOptions2JSON}
\alias{qtlOptions2JSON}
\title{qtlOptions2JSON}
\usage{
qtlOptions2JSON(path = file.path(getwd(), "methQTL_options.json"))
}
\arguments{
\item{path}{A filename, to which the option setting is to be saved}
}
\description{
This function stores the current options setting as a JSON file at the specified path
}
\examples{
{
  qtlSetOption('cluster.cor.threshold'=0.5)
  qtlOptions2JSON("my_opts.json")
  qtlJSON2options("my_opts.json")
}
}
\author{
Michael Scherer
}
