% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtlPlotAnnotationEnrichment}
\alias{qtlPlotAnnotationEnrichment}
\title{qtlPlotAnnotationEnrichment}
\usage{
qtlPlotAnnotationEnrichment(meth.qtl.res, ...)
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{methQTLResult-class}} or a list of such objects.}

\item{...}{Further parameters passed to \code{\link{qtlAnnotationEnrichment}}}
}
\description{
This functions returns and enrichment plot for different genomic annotation enrichments
}
\seealso{
qtl.annotation.enrichment
}
\author{
Michael Scherer
}
