% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtlPlotBaseSubstitution}
\alias{qtlPlotBaseSubstitution}
\title{qtlPlotBaseSubstitution}
\usage{
qtlPlotBaseSubstitution(meth.qtl.res, ...)
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{methQTLResult-class}} or a list of such objects.}

\item{...}{Further parameters passed to \code{\link{qtlBaseSubstitutionEnrichment}}}
}
\value{
None
}
\description{
This function returns an enrichment plot for the different base substitutions.
}
\examples{
{
 meth.qtl.res <- loadMethQTLResult(system.file("extdata/methQTLResult_chr18",package="MAGAR"))
 qtlPlotBaseSubstitution(meth.qtl.res)
}
}
\seealso{
qtlBaseSubstitutionEnrichment
}
\author{
Michael Scherer
}
