% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtlUpSetPlotTagCpGs}
\alias{qtlUpSetPlotTagCpGs}
\title{qtlUpSetPlotTagCpGs}
\usage{
qtlUpSetPlotTagCpGs(meth.qtl.res.list, ...)
}
\arguments{
\item{meth.qtl.res.list}{A list of \code{\link{methQTLResult}} objects, for which correlation blocks are to be overlapped}

\item{...}{Further argument passed to \code{\link[UpSetR]{upset}}}
}
\description{
This function overlaps the tagCpGs for a list of methQTL results
}
\details{
This function draws an UpSetPlot for the overlaps directly from to the open graphics device
}
\author{
Michael Scherer
}
