% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLInput-class.R
\docType{methods}
\name{save.methQTL,methQTLInput-method}
\alias{save.methQTL,methQTLInput-method}
\alias{svae.methQTL,methQTL-method}
\title{save.methQTL}
\usage{
\S4method{save.methQTL}{methQTLInput}(object, path)
}
\arguments{
\item{object}{The \code{\link{methQTLInput-class}} object to be stored on disk.}

\item{path}{A path to a non-existing directory for files to be stored.}
}
\description{
This functions stores a methQTLInput object in disk.
}
\author{
Michael Scherer
}
