% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLResult-class.R
\docType{methods}
\name{save.methQTLResult,methQTLResult-method}
\alias{save.methQTLResult,methQTLResult-method}
\alias{save.methQTLResult,methQTL-method}
\title{save.methQTLResult}
\usage{
\S4method{save.methQTLResult}{methQTLResult}(object, path)
}
\arguments{
\item{object}{The \code{\link{methQTLResult-class}} object to be stored on disk.}

\item{path}{A path to a non-existing directory for files to be stored.}
}
\description{
This functions stores a methQTLInputResult object in disk.
}
\author{
Michael Scherer
}
