% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLInput-class.R
\docType{methods}
\name{saveMethQTL,methQTLInput-method}
\alias{saveMethQTL,methQTLInput-method}
\alias{saveMethQTL,methQTL-method}
\alias{saveMethQTL}
\title{saveMethQTL}
\usage{
\S4method{saveMethQTL}{methQTLInput}(object, path)
}
\arguments{
\item{object}{The \code{\link{methQTLInput-class}} object to be stored on disk.}

\item{path}{A path to a non-existing directory for files to be stored.}
}
\value{
None
}
\description{
This functions stores a methQTLInput object in disk.
}
\examples{
meth.qtl <- loadMethQTL(system.file("extdata","reduced_methQTL",package="MAGAR"))
saveMethQTL(meth.qtl,"methQTLInput")
}
\author{
Michael Scherer
}
