% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQTLInput-class.R
\docType{methods}
\name{saveMethQTLInput,MethQTLInput-method}
\alias{saveMethQTLInput,MethQTLInput-method}
\alias{saveMethQTLInput,methQTL-method}
\alias{saveMethQTLInput}
\title{saveMethQTLInput}
\usage{
\S4method{saveMethQTLInput}{MethQTLInput}(object, path)
}
\arguments{
\item{object}{The \code{\link{MethQTLInput-class}} object to be stored on disk.}

\item{path}{A path to a non-existing directory for files to be stored.}
}
\value{
None
}
\description{
This functions stores a MethQTLInput object in disk.
}
\examples{
meth.qtl <- loadMethQTLInput(system.file("extdata","reduced_methQTL",package="MAGAR"))
saveMethQTLInput(meth.qtl,"MethQTLInput")
}
\author{
Michael Scherer
}
