% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{do.geno.import.idat}
\alias{do.geno.import.idat}
\title{do.geno.import.idat}
\usage{
do.geno.import.idat(
  idat.files,
  s.anno,
  s.id.col,
  out.dir,
  idat.platform = "humanomni258v1a",
  call.method = "krlmm",
  gender.col = NULL
)
}
\arguments{
\item{idat.files}{A path to the directory where the IDAT files are stored}

\item{s.anno}{The sample annotation sheet as a \code{'data.frame'}}

\item{s.id.col}{The column in the sample annotation sheet specifying the sample identifiers}

\item{out.dir}{The output directory}

\item{idat.platform}{The array platform to be used. Check those available using the crlmm function \code{\link{validCdfNames}}.
Additionally, for the Illumina OmniExpress 12 v1.0, and Infinium Omni2.5-8 v1.4 we
provide two custom annotations, which can be used using the 'OmniExpress' and
'OmniExome' options, respectively.}

\item{call.method}{The genotype calling method passed to \code{\link{genotype.Illumina}}}
}
\value{
A vector with three elements: \describe{
  \item{\code{"bed.file"}}{Path to the BED file for PLINK}
  \item{\code{"bim.file"}}{Path to the BIM file for PLINK}
  \item{\code{"fam.file"}}{Path to the FAM file for PLINK}
}
}
\description{
This function imports genotyping data from IDAT files using the \code{'crlmm'} R-package
}
\author{
Michael Scherer
}
