% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLResult-class.R
\docType{methods}
\name{getCorrelationBlocks,methQTLResult-method}
\alias{getCorrelationBlocks,methQTLResult-method}
\title{getCorrelationBlocks}
\usage{
\S4method{getCorrelationBlocks}{methQTLResult}(object, meth.qtl = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{methQTLResult-class}}.}

\item{meth.qtl}{An optional argument of class \code{\link{methQTLInput-class}} comprising CpG annotations
to be used to name the entries in the list.}
}
\value{
A \code{list} object containing the correlation blocks.


}
\description{
Returns the correlation blocks defined for the given dataset
}
