% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLInput-class.R
\docType{methods}
\name{getMethData,methQTLInput-method}
\alias{getMethData,methQTLInput-method}
\alias{getMethData,methQTL-method}
\title{getMethData}
\usage{
\S4method{getMethData}{methQTLInput}(object, site = NULL, sample = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{methQTLInput-class}}.}

\item{site}{The sites to be selected either as a numeric or logical vector. If \code{NULL} all sites are returned.}

\item{sample}{The samples to be selected either as a numeric or logical vector. If \code{NULL} all samples are returned.}
}
\value{
The methylation matrix either as a matrix of \code{\link{HDF5Matrix}}.
}
\description{
Returns methylation information for the given dataset.
}
