% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLInput-class.R
\docType{methods}
\name{getGeno,methQTLInput-method}
\alias{getGeno,methQTLInput-method}
\alias{getPheno,methQTL-method}
\title{getPheno}
\usage{
\S4method{getGeno}{methQTLInput}(object)
}
\arguments{
\item{object}{An object of class \code{\link{methQTLInput-class}}.}
}
\value{
The phenotypic data either as a \code{data.frame}.
}
\description{
Returns phenotypic information for the given dataset.
}
