% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQTLInput-class.R
\docType{methods}
\name{getPheno,MethQTLInput-method}
\alias{getPheno,MethQTLInput-method}
\alias{getPheno,methQTL-method}
\alias{getPheno}
\title{getPheno}
\usage{
\S4method{getPheno}{MethQTLInput}(object)
}
\arguments{
\item{object}{An object of class \code{\link{MethQTLInput-class}}.}
}
\value{
The phenotypic data either as a \code{data.frame}.
}
\description{
Returns phenotypic information for the given dataset.
}
\examples{
meth.qtl <- loadMethQTLInput(system.file("extdata","reduced_methQTL",package="MAGAR"))
head(getPheno(meth.qtl))
}
