% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLResult-class.R
\docType{methods}
\name{getResult,methQTLResult-method}
\alias{getResult,methQTLResult-method}
\title{getResult}
\usage{
\S4method{getResult}{methQTLResult}(object)
}
\arguments{
\item{object}{An of type \code{\link{methQTLResult-class}}.}
}
\value{
The methQTL results as a \code{data.frame} with each row being a methQTL.
}
\description{
Returns the methQTL results stores in the object.
}
