% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLResult-class.R
\docType{methods}
\name{getResult,methQTLResult-method}
\alias{getResult,methQTLResult-method}
\alias{getResult}
\title{getResult}
\usage{
\S4method{getResult}{methQTLResult}(object, cor.blocks = NULL, na.rm = FALSE)
}
\arguments{
\item{object}{An of type \code{\link{methQTLResult-class}}.}

\item{cor.blocks}{Correlation blocks as obtained using \code{getCorrelationBlocks}. Please note that the
correlation blocks need to contain the CpG identifiers, so the \code{\link{methQTLInput-class}} object
needs to be provided to \code{getCorrelationBlocks}.}

\item{na.rm}{Flag indicating if rows containing NA values are to be removed from the result.}
}
\value{
The methQTL results as a \code{data.frame} with each row being a methQTL.
}
\description{
Returns the methQTL results stores in the object.
}
\examples{
meth.qtl.res <- loadMethQTLResult(system.file("extdata","methQTLResult_chr18",package="MAGAR"))
head(getResult(meth.qtl.res))
}
