% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLResult-class.R
\name{loadMethQTLResult}
\alias{loadMethQTLResult}
\title{loadMethQTLResult}
\usage{
loadMethQTLResult(path)
}
\arguments{
\item{path}{Path to the directory that has been created by \code{save.methQTLResult,methQTLInput-method}.}
}
\value{
The object of type \code{\link{methQTLResult-class}} that has been stored on disk.
}
\description{
This functions load a \code{\link{methQTLResult-class}} object from disk.
}
\author{
Michael Scherer
}
