% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{overlapInputs}
\alias{overlapInputs}
\title{overlapInputs}
\usage{
overlapInputs(meth.qtl.list, type)
}
\arguments{
\item{meth.qtl.list}{A list of \code{\link{methQTLInput-class}} or \code{\link{methQTLResult-class}} objects to be overlapped}

\item{type}{The type of annotation to be overlapped. Needs to be \code{'SNP'}, \code{'CpG'} or \code{'cor.block'}}
}
\value{
A data frame containing the annotations of the unique input values.
}
\description{
Overlaps the input annotations
}
\author{
Michael Scherer
}
