% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{qtl.setOption}
\alias{qtl.setOption}
\title{qtl.setOption}
\usage{

  qtl.setOption(rnbeads.options = system.file("extdata/rnbeads_options.xml",
  package = "methQTL"), meth.data.type = "idat.dir",
  rnbeads.report = "temp", rnbeads.qc = F, HDF5dump = F,
  hardy.weinberg.p = 0.001, minor.allele.frequency = 0.05,
  missing.values.samples = 0.05, plink.path = system.file("bin/plink",
  package = "methQTL"), cluster.cor.threshold = 0.2,
  standard.deviation.gauss = 1000, absolute.distance.cutoff = 1e+06,
  linear.model.type = "classial.linear",
  representative.CpG.computation = "majority.median")
}
\arguments{
\item{rnbeads.options}{Path to an XML file specifying the RnBeads options used for data import. The default options
are suitable for Illumina Beads Array data sets.}

\item{meth.data.type}{Type of DNA methylation data used. Choices are listed in \code{\link{rnb.execute.import}}.}

\item{rnbeads.report}{Path to an existing directory, in which the preprocessing report of RnBeads is to be stored.
Defaults to the temporary file.}

\item{rnbeads.qc}{Flag indicating if the quality control module of RnBeads is to be executed.}

\item{HDF5dump}{Flag indicating, if large matrices are to be stored on disk rather than in main memory using the
\code{\link{HDF5Array}} package.}

\item{hardy.weinberg.p}{P-value used for the markers to be excluded if they do not follow the
Hardy-Weinberg equilibrium as implemented in \code{PLINK}.}

\item{minor.allele.frequency}{Threshold for the minor allele frequency of the SNPs to be used in the analysis.}

\item{missing.values.samples}{Threshold specifying how much missing values per SNP are allowed across the samples
to be included in the analyis.}

\item{plink.path}{Path to an installation of PLINK (also comes with the package)}

\item{cluster.cor.threshold}{Threshold for CpG methylatin state correlation to be considered as connected in
the distance graph used to compute the correlation clustering.}

\item{standard.deviation.gauss}{Standard deviation of the Gauss distribution used to weight the correlation
according to its distance.}

\item{absolute.distance.cutoff}{Distance cutoff after which a CpG correlation is not considered anymore.}

\item{linear.model.type}{Linear model type to be used. Can be either \code{"categorical.anova"} or \code{"classical.linear"}.
see \code{\link{call.methQTL.block}} for more informations.}

\item{representative.CpG.computation}{Method how the representative per CpG correlation block is computed. Available
options are \code{"majority.median"} for the site that is the median in most of the samples within the
correlation block, \code{"mean.center"} for an artifical site in the geometric center of the block with
the average methylation level or \code{"best.all"} for the CpG with the best p-value across all of the
CpGs in the correlation block.}
}
\description{
Change global options for methQTL calculation
}
\examples{
\donttest{
qtl.getOption("rnbeads.report")
qtl.setOption(rnbeads.report=getwd())
qtl.getOption("rnbeads.report")
}
}
\author{
Michael Scherer
}
