% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtlPlotClusterSize}
\alias{qtlPlotClusterSize}
\title{qtlPlotClusterSize}
\usage{
qtlPlotClusterSize(meth.qtl.res, type = "count")
}
\arguments{
\item{meth.qtl.res}{An object of type \code{\link{MethQTLResult-class}}}

\item{type}{Either \code{"genomic"} or \code{"count"}, for genomic size of the correlation
block in base pairs or as the number of CpGs}
}
\value{
An object of type ggplot containing the histogram as a plot
}
\description{
This functions returns a histogram comprising the (genomic) sizes of the correlation blocks
in the given objet.
}
\examples{
meth.qtl.res <- loadMethQTLResult(system.file("extdata","MethQTLResult_chr18",package="MAGAR"))
qtlPlotClusterSize(meth.qtl.res)
}
\author{
Michael Scherer
}
