% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{qtlUpSetPlotTagCpGs}
\alias{qtlUpSetPlotTagCpGs}
\title{qtlUpSetPlotTagCpGs}
\usage{
qtlUpSetPlotTagCpGs(meth.qtl.res.list, ...)
}
\arguments{
\item{meth.qtl.res.list}{A list of \code{\link{MethQTLResult-class}} objects, for which correlation blocks are to be overlapped}

\item{...}{Further argument passed to \code{\link[UpSetR]{upset}}}
}
\value{
None
}
\description{
This function overlaps the tagCpGs for a list of methQTL results
}
\details{
This function draws an UpSetPlot for the overlaps directly from to the open graphics device
}
\examples{
meth.qtl.res.1 <- loadMethQTLResult(system.file("extdata","MethQTLResult_chr18",package="MAGAR"))
meth.qtl.res.2 <- meth.qtl.res.1
qtlUpSetPlotTagCpGs(list(A=meth.qtl.res.1,B=meth.qtl.res.2))
}
\author{
Michael Scherer
}
