% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methQTLResult-class.R
\docType{methods}
\name{saveMethQTLResult,methQTLResult-method}
\alias{saveMethQTLResult,methQTLResult-method}
\alias{saveMethQTLResult,methQTL-method}
\alias{saveMethQTLResult}
\title{saveMethQTLResult}
\usage{
\S4method{saveMethQTLResult}{methQTLResult}(object, path)
}
\arguments{
\item{object}{The \code{\link{methQTLResult-class}} object to be stored on disk.}

\item{path}{A path to a non-existing directory for files to be stored.}
}
\value{
None
}
\description{
This functions stores a methQTLInputResult object in disk.
}
\examples{
{
 meth.qtl.res <- loadMethQTLResult(system.file("extdata/methQTLResult_chr18",package="MAGAR"))
 saveMethQTLResult(meth.qtl.res,"methQTLResult")
}
}
\author{
Michael Scherer
}
