% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethQTLResult-class.R
\docType{methods}
\name{saveMethQTLResult,MethQTLResult-method}
\alias{saveMethQTLResult,MethQTLResult-method}
\alias{saveMethQTLResult,methQTL-method}
\alias{saveMethQTLResult}
\title{saveMethQTLResult}
\usage{
\S4method{saveMethQTLResult}{MethQTLResult}(object, path)
}
\arguments{
\item{object}{The \code{\link{MethQTLResult-class}} object to be stored on disk.}

\item{path}{A path to a non-existing directory for files to be stored.}
}
\value{
None
}
\description{
This functions stores a MethQTLInput object in disk.
}
\examples{
meth.qtl.res <- loadMethQTLResult(system.file("extdata","MethQTLResult_chr18",package="MAGAR"))
saveMethQTLResult(meth.qtl.res,"MethQTLResult")
}
\author{
Michael Scherer
}
